requirements.txt
streamlit
openai
mysql-connector-python
pandas
python-dotenv
sqlalchemy
google-generativeai